vilegal/requirements.txt
omegaconf==2.0.6
pytorch-lightning==1.1.7
hydra-core==1.0.6
transformers==4.20.1
datasets==1.3.0
rouge-score==0.0.4
torch>=1.9.0
torchvision
torchaudio
psutil==5.8.0
scikit-learn
numpy
pandas 